sytle(*): first implementation
