+++ lib/config/config.json
@@ -1,0 +1,3 @@
+{
+  "appName": "Lyra supercharged 💪"
+}
